leds: leds-qpnp: make blink functionality synchronous
pwm_cfg->blinking is a shared parameter between qpnp_mpp_set
and led_blink function. When a blink operation is given to
the led when it is glowing with some brightness value, the led
does not blink because the blinking parameter is wrongly
reset to false in the end of qpnp_mpp_set. So making the blink
functionality synchronous will solve the issue.

Change-Id: Ie84fe52bc84d5c14573e3f1a042e163e62152cce
Signed-off-by: Abinaya P <abinayap@codeaurora.org>
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>